✨ feat: revamp settings screen with account and subscription management
The commit reorganizes the settings screens by:
- Merging profile and subscription screens into a unified account screen
- Adding subscription management with RevenueCat integration
- Improving navigation flow and user experience
- Updating profile menu to reflect new structure
